Explaining to commit 95a9ffac17b44641e960126f56d73adaaef7b415
@ProgVal has failed to give correct commands at https://github.com/ProgVal/Supybot-docs/blob/master/i18n/Limnoria_i18n.tex#L181
and then complains that everyone typo with that command. He is complaining that
I have broken the messages.pot file with his incorrect command, but he forgets
that he has given that incorrect command. He is also using different versions
of programs which we should use and doesn't tell that to us.