Eliminate unnecessary closing of the log file.
Name returns a string that is the log file name: the channel plus the timestamp.
But on my system log.name returns the fully qualified path to the currently open file. This is because log is a file handle, so if you declare it with a path, it will be included in the name.
On systems were logging is not in the current directory, the two will never match, and the file is closed and then reopened.
Only taking the basename solves this problem.