src/conf.py: default reply.error.detailed to True.
Most of people are just pasting the `An error has occurred and has been logged. Check the logs for more informations.` message without even looking at the logs and they are then told to set this as True.

If this defaulted to True, we would probably tell them to look at messages.log anyway, but this might help. Any opinions?