.travis.runonce.bash: care about exit status
This excludes msgcheck, because it's over-sensitive and always returns
exit value which logic I don't remember unless everything is correct
according to it.

Oh and the behaviour before this commit allows everything to fail
without having exit code which shows that it failed and we wouldn't
notice any failures in near future.