src: update help text of commands.nested.brackets
The shift part only applied to people on American keyboard and not all
users are on that, so it only caused confusion.

Strong historical motivations should be enough for most of people as I
wasn't able to remember the correct brackets when I tried changing them
to `<>` which I feel are the easiest to type on Finnish/Swedish keyboard.

Also update messages.pot & l10n-fi.