Clearsign authorisation via subkey.
Replacement code which might work to enable advanced keys with signing
subkeys to be correctly handled by the bot by associating the subkey
with the relevant master key.

Signing format still only clearsigning, the key details are more
important and auth via encrypted token and decryption is likely to be
more reliable anyway as there is far less chance of some other protocol
messing with the signed content.  Effectively no chance, though the odd
corrupted packet here and there is still possible.  Whereas with
clearsigning it can be broken by all manner of rewriting in
transit (which happens often enough with signed email as it is).

See also Issue #1045 for greater detail of what needs to be fixed and
what is to be done about it.