debian/compat is required, v9 or above
DebHelper, which helps build Debian packages, fails for the Debian packaging as is.  This is due to the system autoassuming a compat level of Debhelper 1, and anything older than v4 is not accepted.  Debian Testing, Ubuntu Xenial, etc. have DebHelper compat 9 as the recommended (this is also seen in the Debian Maintainer Guide (on August 6, 2016): https://www.debian.org/doc/manuals/maint-guide/dother.en.html#compat)