Use a monotonic time for registry cache.
So an old cache does not take precedence over a newly set value.
I noticed this bug because of the time going backward in tests
because of the time.time mock introduced in
dcf55cf6deb90f2498e53068175562aa36d315cc, but this may happen
in production systems too.

Also adds another deprecation warning for python < 3.3.