callbacks: Deduplicate computation of message overhead.
NestedCommandsIrcProxy.reply tried to compute the length of
a message's overhead (':<prefix> <command> <target> :' and '\r\n'),
but because it doesn't exactly duplicate the code of _makeReply --
especially when computing the actual target -- it ends up slightly
off.

This commit moves the overhead length computation next to the
computation of the overhead itself, so it's always exact.

This will also allow message splitting in other methods without
duplicating the code too much.