Log current version on startup
We recently got a bug report from someone running Limnoria in Docker, and it turned
out that a plugin's import was segfaulting and Docker silently restarted the process.

I wasted some time figuring out what happened because I didn't notice the process
was being restarted. This should make this more noticeable in the future.