NickAuth: refer to "services accounts" instead of "nicks" in help
On modern networks, account names are usually separated from your current nick,
so the idea of registering "nicks" is confusing. For instance, you can be using
a grouped nick that is not your main account name, and WHOIS will only ever
show your main account name.

Also, the bot username is optional in the `add` and `remove` commands - update
the docs to reflect that.